refactor: move request component to components/request
